Modularise GearCooldownTimer
